# docs/swagger.yaml
basePath: /gradiol-api/api
definitions:
...
  dto.CreateProjectReq:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      workspace_id:
        type: string
    required:
    - name
    - workspace_id
    type: object
  dto.CreateWorkspaceReq:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  dto.UpdateProjectReq:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
  dto.UpdateWorkspaceReq:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
  model.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: string
    type: object
...
  model.Workspace:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  pkg.AppError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
host: asia-southeast1-renzip-478811.cloudfunctions.net
info:
  contact: {}
  description: Backend API for Graphic Diagram Online (GraDiOl)
  title: GraDiOl API
  version: "1.0"
paths:
...
  /projects:
    post:
      consumes:
      - application/json
      description: Creates a new project in a workspace
      parameters:
      - description: Project Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - projects
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a project and its documents
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates a project's details
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - projects
...
  /workspaces:
    get:
      consumes:
      - application/json
      description: Retrieves all workspaces the current user is a member of
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Get list of workspaces
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      description: Creates a new workspace and sets the current user as owner
      parameters:
      - description: Workspace Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWorkspaceReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Workspace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Create a new workspace
      tags:
      - workspaces
  /workspaces/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a workspace and all its contents (must be owner)
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Delete a workspace
      tags:
      - workspaces
    put:
      consumes:
      - application/json
      description: Updates workspace details (must be owner or editor)
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWorkspaceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Workspace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Update a workspace
      tags:
      - workspaces
  /workspaces/{id}/projects:
    get:
      consumes:
      - application/json
      description: Retrieves all projects for a given workspace ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Get list of projects in a workspace
      tags:
      - projects
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"